requirements.txt
# Core Dependencies
python-dotenv==1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
colorama>=0.4.6
tqdm>=4.65.0

# AI/ML Frameworks
torch>=2.0.0
transformers>=4.30.0
langchain>=0.1.0
autogen>=0.2.0
sentence-transformers>=2.2.0

# Financial Data
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0

# Vector Database
chromadb>=0.4.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0